添加热修复的代码

+++ buildSrc/src/main/java/com/wuliqinwang/patch/PatchPlugin.kt
@@ -10,6 +10,7 @@ import org.apache.commons.compress.utils.IOUtils
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.Task
+import org.gradle.api.tasks.TaskInputs
 import org.objectweb.asm.Opcodes
 import java.io.File
 import java.io.FileInputStream
@@ -28,15 +29,6 @@ class PatchPlugin : Plugin<Project> {
     companion object {
         // DES: 混淆文件名
         private const val MAPPING_FILE_NAME = "mapping.txt"
-
-        // DES: 用于记录需要热修复的Class的md5值
-        private const val MD5_FILE_NAME = "classesMd5.txt"
-
-        // DES: 用于存放需要热修复的Class文件
-        private const val PATCH_CLASS_FILE_NAME = "patchClasses.jar"
-
-        // DES: 热修复包文件名
-        private const val PATCH_DEX_FILE_NAME = "patch.dex"
     }
 
     override fun apply(target: Project) {
@@ -44,17 +36,17 @@ class PatchPlugin : Plugin<Project> {
             throw RuntimeException("非app应用，不适合该插件")
         }
         // DES: 在build.gradle文件中创建一个热修复配置字段
-        target.extensions.create(HotFixPatch.HOT_FIX_PATCH, HotFixPatch::class.java)
+        target.extensions.create(HotFixExtension.HOT_FIX_PATCH, HotFixExtension::class.java)
         // DES: 在build.gradle解析完成后执行我们自己的
         target.afterEvaluate { project ->
             // DES: 获取热修复的配置信息
-            val hotFixPatch = project.extensions.getByType(HotFixPatch::class.java)
+            val hotFixPatch = project.extensions.getByType(HotFixExtension::class.java)
             // DES: 获取app的配置信息
             val appExtension = project.extensions.findByType(AppExtension::class.java)
             // DES: 变量应用程序的变体，比如debug或者release环境等
             appExtension?.applicationVariants?.all { variant ->
                 // DES: 如果实在debug环境，并且热修复在debug的开发是关闭的，则不进行热修复包的生成
-                if (variant.name.contains("debug") && !hotFixPatch.isDebugOn) {
+                if (variant.name.contains("debug") && !hotFixPatch.debugOn) {
                     return@all
                 }
                 // DES: 生成热修复包以及其他相关文件
@@ -65,35 +57,44 @@ class PatchPlugin : Plugin<Project> {
 
     private fun generatePatchDexFile(
         project: Project,
-        hotFixPatch: HotFixPatch,
+        hotFixPatch: HotFixExtension,
         variant: ApplicationVariant
     ) {
         val variantName = PatchUtils.getCapitalLetter(
             variant.name
         ) ?: "debug"
         // DES: 获取patch的包存放文件目录
-        val patchOutputDir = getPatchOutputDir(project, hotFixPatch, variantName)
+        val patchOutputDir = getPatchOutputDir(project, hotFixPatch, variant.name)
         // DES: 获取Android的混淆任务
         val proguardTask = project.tasks
             .findByName("minify${variantName}WithR8")
         // DES: 获取备份的混淆映射文件
-        val mappingBackFile = File(patchOutputDir,
-            MAPPING_FILE_NAME
-        )
+        val mappingBackFile = File(patchOutputDir, MAPPING_FILE_NAME)
         // DES: 开启混淆，我则在混淆完成后，进行备份混淆的映射文件
-        backupMappingFile(project, proguardTask, mappingBackFile)
+        backupMappingFile(
+            project,
+            proguardTask,
+            mappingBackFile
+        ) { taskInputs ->
+            doPatchDexFile(project, taskInputs, patchOutputDir, hotFixPatch, variant)
+        }
         // DES: 添加之前的混淆文件
         applyTestedMapping(proguardTask as? R8Task, mappingBackFile)
-        // DES: 这里就开始执行生成补丁包文件
-        doPatchDexFile(project, patchOutputDir, hotFixPatch, variant, variantName)
     }
 
     // DES: 执行备份混淆文件到指定文件
-    private fun backupMappingFile(project: Project, proguardTask: Task?, mappingBackFile: File) {
-        proguardTask?.doLast {
+    private fun backupMappingFile(
+        project: Project,
+        proguardTask: Task?,
+        mappingBackFile: File,
+        action: (TaskInputs) -> Unit
+    ) {
+        proguardTask?.doFirst {
+            println("start minifyDebugWithR8===================")
             for (file in it.outputs.files.files) {
                 if (file.name.endsWith(MAPPING_FILE_NAME)) {
                     try {
+                        if (!file.exists()) break
                         FileUtils.copyFile(file, mappingBackFile)
                         project.logger.debug("备份混淆文件到: ${mappingBackFile.canonicalPath}")
                     } catch (e: Exception) {
@@ -102,6 +103,7 @@ class PatchPlugin : Plugin<Project> {
                     break
                 }
             }
+            action(it.inputs)
         }
     }
 
@@ -111,14 +113,16 @@ class PatchPlugin : Plugin<Project> {
             proguardTask.testedMappingFile.apply {
                 var isOldMappingFile = false
                 for (file in files) {
+                    println("file name=${file.name}")
                     if (file.name == mappingFile.name) {
                         isOldMappingFile = true
                         break
                     }
                 }
                 if (!isOldMappingFile) {
-                    files.add(mappingFile)
+                    setFrom(mappingFile)
                 }
+                println("files size=${files.size} files type=${files::class.java}")
             }
         }
     }
@@ -126,63 +130,86 @@ class PatchPlugin : Plugin<Project> {
     // DES: 得到差分包的输出目录
     private fun getPatchOutputDir(
         project: Project,
-        hotFixPatch: HotFixPatch,
+        hotFixPatch: HotFixExtension,
         variantName: String
     ): File {
         return (if (hotFixPatch.patchOutputDir.isNotEmpty()) {
             File(hotFixPatch.patchOutputDir, variantName)
         } else {
             File(project.buildDir, "patch/${variantName}")
-        }).apply {
-            if (!exists()) {
-                mkdirs()
+        }).let {
+            if (!it.exists()) {
+                it.mkdirs()
             }
+            it
         }
     }
 
     // DES: 做生成补丁包的开始方法
     private fun doPatchDexFile(
         project: Project,
+        taskInputs: TaskInputs,
         outputDir: File,
-        hotFixPatch: HotFixPatch,
-        variant: ApplicationVariant,
-        variantName: String
+        hotFixPatch: HotFixExtension,
+        variant: ApplicationVariant
     ) {
-        project.tasks.findByName("compile${variantName}JavaWithJavac")?.doLast {
-            val generator =
-                PatchGenerator(project, outputDir)
-            val classFiles = it.outputs.files.files
-            val applicationName = hotFixPatch.applicationName.replace(
-                "\\.",
-                Matcher.quoteReplacement(File.separator)
-            )
-            val md5Map = HashMap<String, String>(classFiles.size)
-            for (file in classFiles) {
-                val filePath = file.absolutePath
-                when {
-                    filePath.endsWith(".jar") -> {
-                        processJar(
-                            applicationName,
-                            file
-                        ) { className, md5Hex, codeByteArray ->
-                            md5Map[className] = md5Hex
-                            generator.addClassToJarFile(className, md5Hex, codeByteArray)
-                        }
-                    }
-                    filePath.endsWith(".class") -> {
-                        processClass(
-                            applicationName,
-                            variant.dirName,
-                            file
-                        ) { className, md5Hex, codeByteArray ->
-                            md5Map[className] = md5Hex
-                            generator.addClassToJarFile(className, md5Hex, codeByteArray)
-                        }
-                    }
+        println("start for each classes===================")
+        val generator = PatchGenerator(project, outputDir)
+        val classFiles = taskInputs.files.files
+        val applicationName = hotFixPatch.applicationName.replace(
+            "\\.",
+            Matcher.quoteReplacement(File.separator)
+        )
+        val md5Map = HashMap<String, String>(classFiles.size)
+        for (file in classFiles) {
+            fileForEach(file) { targetFile ->
+                handleClassAndJarFile(
+                    targetFile,
+                    applicationName,
+                    variant.dirName
+                ) { className, md5Hex, codeByteArray ->
+                    md5Map[className] = md5Hex
+                    generator.addClassToJarFile(className, md5Hex, codeByteArray)
                 }
             }
-            // DES: 在这里生成热修复包
-            generator.generate(md5Map)
+        }
+        // DES: 在这里生成热修复包
+        generator.generate(md5Map)
+    }
+
+    // DES: 循环遍历文件
+    private fun fileForEach(targetFile: File?, action: (file: File) -> Unit) {
+        if (targetFile == null || targetFile.isFile) {
+            targetFile?.apply {
+                action(this)
+            }
+        } else {
+            val listFiles = targetFile.listFiles()
+            if (!listFiles.isNullOrEmpty()) {
+                listFiles.forEach {
+                    fileForEach(it, action)
+                }
+            } else {
+                action(targetFile)
+            }
+        }
+    }
+
+    // DES: 处理Class和jar文件
+    private inline fun handleClassAndJarFile(
+        file: File,
+        applicationName: String,
+        dirName: String,
+        callback: (String, String, ByteArray) -> Unit
+    ) {
+        val filePath = file.absolutePath
+        when {
+            filePath.endsWith(".jar") -> {
+//                processJar(applicationName, file, callback)
+            }
+            filePath.endsWith(".class") -> {
+                processClass(applicationName, dirName, file, callback)
+            }
         }
     }
 
@@ -194,6 +221,7 @@ class PatchPlugin : Plugin<Project> {
         handleCallback: (String, String, ByteArray) -> Unit
     ) {
         val className = file.absolutePath.split(dirName)[1].substring(1)
+        println(className)
         if (!isClassHandle(applicationName, className)) {
             return
         }
@@ -223,7 +251,8 @@ class PatchPlugin : Plugin<Project> {
     ) {
         val tempApplicationName = applicationName.replace(
             Matcher.quoteReplacement(File.separator),
-            "/")
+            "/"
+        )
         val backupFile = File(file.parent, "${file.name}.bak")
         // DES: 用于临时缓存
         JarOutputStream(FileOutputStream(backupFile)).use { jarOs ->
@@ -278,7 +307,7 @@ class PatchPlugin : Plugin<Project> {
                     override fun visitInsn(opCodes: Int) {
                         // DES: 在构造器中添加一个FixCompatible类的引用
                         if (name == "<init>" && opCodes == Opcodes.RETURN) {
-                            super.visitLdcInsn(Type.getType("Lcom.wuliqinwang.hotfix.FixCompatible;"))
+                            super.visitLdcInsn(Type.getType("Lcom/wuliqinwang/hotfix/FixCompatible;"))
                         }
                         super.visitInsn(opCodes)
                     }
